+++ demos/world/src/main/java/is/codion/demos/world/ui/CountryEditPanel.java
@@ -85,7 +85,7 @@ final class CountryEditPanel extends EntityEditPanel {
 		NumberField<Double> averageCityPopulationField = doubleField()
 						.link(editModel.averageCityPopulation())
 						.maximumFractionDigits(2)
-						.groupingUsed(true)
+						.grouping(true)
 						.horizontalAlignment(SwingConstants.CENTER)
 						.focusable(false)
 						.editable(false)